Fixes cached client.
The neuprint python functions fetch_all_rois and fetch_custom will use
a cached version of the client, if not explicitly set in the function
call. This results in the first query to the server setting the dataset
and all future queries using that cached datset, even when a new dataset
is passed. Adding a new client to each function call makes the cache
update between each request.